Update PsxCounters with Mednafen-derived logic and GPU_vBlank support
- Ported root counter logic from Mednafen-PC3SX.
- Updated `PsxCounters.h` to use 16-bit `mode` and `target` members.
- Updated PPC recompiler (`pR3000A.c`) to handle 16-bit counter registers using `LHZ`.
- Added support for `GPU_vBlank` callback in `plugins.c`, `plugins.h`, and `Plugin.h`.
- Ensured `GPU_vBlank` is initialized with a stub to prevent NULL dereferences.
- Fixed `DebugVSync()` placement to maintain original PS3SX frame behavior.
- Retained `psxUpdateVSyncRate` for compatibility with other core modules.

+++ PS3SX/pcsxcore/PsxCounters.h
@@ -30,7 +30,7 @@ extern "C" {
 #include "plugins.h"
 
 typedef struct {
-    u32 mode, target;
+    u16 mode, target;
     u32 rate, irq, counterState, irqState;
     u32 cycle, cycleStart;
 } psxCounter;
